requirements.txt
flask==3.0.3;
python-dotenv==1.0.1;
requests==2.32.3;
spotipy==2.24.0;